requirements.txt
flask
psycopg2-binary
torch
torchvision
pillow